requirements.txt
Django>=4.2.0,<5.0.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
python-decouple>=3.8
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
django-allauth>=0.54.0
openai>=1.0.0
google-generativeai>=0.3.0
reportlab>=4.0.0
gunicorn>=21.2.0